requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
config==0.5.0.post0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-on-heroku==1.1.2
django-registration==3.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
python-decouple==3.4
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
